# code/backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/users/{userId}:
    put:
      tags:
      - user-rest-controller
      summary: Update the profile of the user
      operationId: editProfile
      parameters:
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      - name: newUser
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/User'
      - name: imageFile
        in: query
        required: true
        schema:
          type: string
          format: binary
      responses:
        "404":
          description: 'Profile not found '
          content: {}
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "403":
          description: forbiden o dont have permissions
...
  /api/users/{userId}/checkout:
    post:
      tags:
      - user-rest-controller
      summary: make a purchase
      operationId: checkoutProcess
      parameters:
      - name: billing_street
        in: query
        required: true
        schema:
          type: string
      - name: billing_apartment
        in: query
        required: true
        schema:
          type: string
      - name: billing_city
        in: query
        required: true
        schema:
          type: string
      - name: billing_country
        in: query
        required: true
        schema:
          type: string
      - name: billing_postcode
        in: query
        required: true
        schema:
          type: string
      - name: billing_phone
        in: query
        required: true
        schema:
          type: string
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: 'Game not found '
          content: {}
        "201":
          description: the purchase has been successfully completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "403":
          description: forbiden o dont have permissions
  /api/users/{userId}/cart/{id}:
    post:
      tags:
      - user-rest-controller
      summary: Add a game to the cart of the user
      operationId: addCart
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: 'Game not found '
          content: {}
        "201":
          description: Game added to the cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "403":
          description: forbiden o dont have permissions
    delete:
      tags:
      - user-rest-controller
      summary: Delete a game to the cart of the user
      operationId: deleteCart
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: 'Game not found '
          content: {}
        "200":
          description: Game deleted to the cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "403":
          description: forbiden o dont have permissions
  /api/users/:
    post:
      tags:
      - user-rest-controller
      summary: Register a new user
      operationId: register
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: lastName
        in: query
        required: true
        schema:
          type: string
      - name: mail
        in: query
        required: true
        schema:
          type: string
      - name: password
        in: query
        required: true
        schema:
          type: string
      - name: aboutMe
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: forbiden o dont have permissions
...
  /api/users/{userId}/purchase/{purchaseId}:
    get:
      tags:
      - user-rest-controller
      summary: Get the purchase of the user
      operationId: purchase
      parameters:
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      - name: purchaseId
        in: path
        description: id of the purchase
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: 'Purchase not found '
          content: {}
        "200":
          description: Found the purchase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        "403":
          description: forbiden o dont have permissions
  /api/users/{userId}/moreCartGames/{page}:
    get:
      tags:
      - user-rest-controller
      summary: Get more cart games of the user
      operationId: getMoreCartGames
      parameters:
      - name: page
        in: path
        description: page of the games
        required: true
        schema:
          type: integer
          format: int32
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Games not found in the cart
          content: {}
        "200":
          description: Found more games in the cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /api/users/{userId}/cart:
    get:
      tags:
      - user-rest-controller
      summary: Get the cart of the user
      operationId: cart
      parameters:
      - name: userId
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Cart not found
          content: {}
        "200":
          description: Found the cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get user
      operationId: getUser
      parameters:
      - name: id
        in: path
        description: id of the user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User not found
          content: {}
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: forbiden o dont have permissions
  /api/users/{id}/imageProfile:
    get:
      tags:
      - user-rest-controller
      summary: Get the image profile of the user
      operationId: downloadImageProfile
      parameters:
      - name: id
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: image profile not found
          content: {}
        "200":
          description: Found the image profile
          content:
            image/png: {}
        "403":
          description: forbiden o dont have permissions
  /api/users/recomended:
    get:
      tags:
      - user-rest-controller
      summary: Get recomendation tothe current user
      operationId: recomendations
      parameters:
      - name: numberOfGames
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not found games recomendation
          content: {}
        "200":
          description: Found games recomendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  /api/users/me:
    get:
      tags:
      - user-rest-controller
      summary: Get information about the current user
      operationId: profile
      responses:
        "404":
          description: 'User not found '
          content: {}
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        "403":
          description: forbiden o dont have permissions
...
  /api/games/{id}/coverImage:
    get:
      tags:
      - game-rest-controller
      summary: Get the cover image of a game
      operationId: downloadImageProfile_1
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Cover not found
          content: {}
        "200":
          description: Found the cover image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mail:
          type: string
        name:
          type: string
        lastName:
          type: string
        aboutMe:
          type: string
        billingInformation:
          type: string
        roles:
          type: array
          items:
            type: string
        profilePircture:
          type: string
        totalPrice:
          type: number
          format: float
    Game:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        os:
          type: string
        processor:
          type: string
        memory:
          type: string
        graphics:
          type: string
        directX:
          type: string
        network:
          type: string
        hardDrive:
          type: string
        soundCard:
          type: string
        deleted:
          type: boolean
        totalRating:
          type: integer
          format: int32
        starDistribution:
          type: array
          items:
            type: integer
            format: int32
        gameplayImages:
          type: array
          items:
            type: string
        titleImage:
          type: string
        reviewCount:
          type: integer
          format: int32
        starDistributionInt:
          type: array
          items:
            type: integer
            format: int32
    UserProfile:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        user:
          $ref: '#/components/schemas/User'
        date:
          type: string
          format: date
...